Add support for dash ('-') in titles & keys
The dash should really be supported as a title/key part. It it a commonly used "spacer" in titles and keys, and AWS INI files already support it.

Partially deals with accense/eini#2, but I have left spaces for another day—it is not clear what the policy should be here.